requirements.txt
Flask 
requests 
APScheduler
python-dotenv
zoneinfo